+++ cif/documentation/generate.js
@@ -14,7 +14,7 @@
 
 'use strict';
 
-const magentoSchema = require('../resources/magento-schema-2.4.2.min.json');
+const magentoSchema = require('../resources/magento-schema-2.4.3.min.json');
 const SchemaPruner = require('./SchemaPruner.js');
 const gitClone = require('git-clone');
 const fs = require('fs');
@@ -68,7 +68,7 @@ function generate() {
           fs.writeFileSync(
             path.join(
               __dirname,
-              '../resources/magento-schema-2.4.2.pruned.json'
+              '../resources/magento-schema-2.4.3.pruned.json'
             ),
             JSON.stringify(prunedSchema, null, 2)
           );
@@ -76,7 +76,6 @@ function generate() {
       );
     }
   );
-
 }
 
 // The file contains multiple single-line queries